Embeds

+++ files/claims/download.py
@@ -1,3 +1,5 @@
+from os.path import getsize, exists
+from os import remove
 import requests
 import sys
 
@@ -7,9 +9,17 @@ def downloadGoogleSheet():
 
         response = requests.get(link)
         if response.status_code == 200:
-            with open(file, 'wb') as f:
-                f.write(response.content)
-                print('File saved to: {}'.format(file))
+            NEW = True # to be overwritten if file has the same size
+            with open(f"_{file}", 'wb') as tf:
+                tf.write(response.content)
+            if exists(file):
+                if getsize(file) == getsize(f"_{file}"):
+                    NEW = False
+            if NEW: # if file differs or doesn't exist
+                with open(file, 'wb') as f:
+                    f.write(response.content)
+                    print('File saved to: {}'.format(file))
+            remove(f"_{file}")
         else:
             print(f'Error downloading Google Sheet: {response.status_code}')
             sys.exit(1)
